📦 Pull tab/grid/card directives into external packages
