+++ .changeset/brown-onions-fix.md
@@ -1,0 +1,5 @@
+---
+'myst-cli': patch
+---
+
+Pull tab/grid/card directives into external packages
