+++ packages/myst-cli/package.json
@@ -74,6 +74,9 @@
     "myst-cli-utils": "^0.0.10",
     "myst-common": "^0.0.12",
     "myst-config": "^0.0.9",
+    "myst-ext-card": "^0.0.1",
+    "myst-ext-grid": "^0.0.1",
+    "myst-ext-tabs": "^0.0.1",
     "myst-frontmatter": "^0.0.8",
     "myst-spec": "^0.0.4",
     "myst-templates": "^0.1.6",
