+++ packages/myst-ext-card/README.md
@@ -1,0 +1,3 @@
+# myst-ext-card
+
+`mystjs` extension for `card` directive
