+++ packages/myst-ext-grid/README.md
@@ -1,0 +1,3 @@
+# myst-ext-grid
+
+`mystjs` extension for `grid` directive
