+++ packages/myst-ext-grid/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "allowJs": true,
+    "target": "es6"
+  },
+  "exclude": []
+}
