+++ packages/myst-ext-tabs/README.md
@@ -1,0 +1,3 @@
+# myst-ext-tabs
+
+`mystjs` extension for `tab-set` and `tab-item` directives
